Normalize section headers in KNN lab...
Google Colab supports section folding, but it only works
when each heading it in it's own "text" block.

Additionally, Google Colab will generate an automatic
"Document Outline," but it skips headings that look like 
they begin lists. The sub-headers "3. *" and "4. *" are 
therefore omitted. Simply removing the . character is
enough to make this feature work properly.

Finally, this also standardizes the formatting 
of the "Discussion" prompts. Whereas they were previously 
inconsistent, they are all now _itialicized_ and marked as
an H4. This causes them to show up in the document outline.